misc/HtmlTestSuite/results/test-0568.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8"/>
    <title>CSS Shapes Tests: Circle - invalid r values</title>
    <link href="mailto:biqing.kwok@qq.com" rel="author" title="biqing"/>
    <link href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes" rel="help"/>
    <link href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle" rel="help"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <meta content="dom" name="flags"/>
    <meta content="shape-outside declaration is invalid if circle shape function has an invalid r parameter" name="assert"/>
    <style type="text/css">
    .circle {
        float: left;
        width: 100px;
        height: 100px;
    }
    #circle1 {
        shape-outside: circle(auto);
    }
    #circle2 {
        shape-outside: circle(inherit);
    }
    #circle3 {
        shape-outside: circle(#FFFFFF);
    }
    </style>
</head>
<body>
    <div id="log"/>

    <div class="circle" id="circle1"/>
    <div class="circle" id="circle2"/>
    <div class="circle" id="circle3"/>

    <script>

    test(function() {
        assert_true(getComputedStyle(document.getElementById('circle1')).shapeOutside == 'none');
        assert_true(getComputedStyle(document.getElementById('circle2')).shapeOutside == 'none');
        assert_true(getComputedStyle(document.getElementById('circle3')).shapeOutside == 'none');
    }, 'Circle - invalid r values');

    </script>
</body>
</html>
